不具合関係のAPIに対処
