+++ backend/apps/bug_note/urls.py
@@ -1,10 +1,8 @@
 from django.urls import path
-from apps.bug_note.views import DefectView, DefectDetailView, DefectUpdateView, DefectDeleteView
+from apps.bug_note.views import DefectView, DefectDetailView
 
 
 urlpatterns = [
     path("defects/", DefectView.as_view(), name="defect"),
     path("defects/<int:pk>/", DefectDetailView.as_view(), name="defect-detail"),
-    path("defects/<int:pk>/update/", DefectUpdateView.as_view(), name="defect-update"),
-    path('defects/<int:pk>/delete/', DefectDeleteView.as_view(), name='defect-delete'),
 ]
