+++ backend/apps/bug_note/views/__init__.py
@@ -1,4 +1,2 @@
 from .defect_view import DefectView
 from .defect_detail_view import DefectDetailView
-from .defect_update_view import DefectUpdateView
-from .defect_delete_view import DefectDeleteView
