+++ backend/apps/bug_note/views/defect_detail_view.py
@@ -5,9 +5,10 @@ from drf_spectacular.utils import extend_schema
 from rest_framework import status
 from django.shortcuts import get_object_or_404
 from apps.bug_note.models.defect import Defect
-from apps.bug_note.serializers import DefectDetailSerializer
+from apps.bug_note.serializers import DefectDetailSerializer, DefectUpdateSerializer
 from apps.staff_hub.permission import HasUserPermissionObject
-from apps.bug_note.common import check_bug_note_view_permission
+from apps.bug_note.common import check_bug_note_view_permission, check_bug_note_edit_permission
+from django.utils import timezone
 
 
 class DefectDetailView(APIView):
@@ -23,3 +24,32 @@ class DefectDetailView(APIView):
         defect = get_object_or_404(Defect, pk=pk, deleted_at__isnull=True)
         serializer = DefectDetailSerializer(defect)
         return Response(serializer.data, status=status.HTTP_200_OK)
+
+    @extend_schema(
+        request=DefectUpdateSerializer,
+        responses={200: DefectUpdateSerializer},
+        tags=["defect"],
+        description="不具合を編集"
+    )
+    def put(self, request, pk):
+        check_bug_note_edit_permission(request)
+
+        defect = get_object_or_404(Defect, pk=pk)
+        serializer = DefectUpdateSerializer(defect, data=request.data)
+        serializer.is_valid(raise_exception=True)
+        serializer.save()
+
+        return Response(serializer.data, status=status.HTTP_200_OK)
+
+    @extend_schema(
+        tags=["defect"],
+        description="不具合を論理削除する",
+        responses={204: None}
+    )
+    def delete(self, request, pk):
+        check_bug_note_edit_permission(request)
+        defect = get_object_or_404(Defect, pk=pk, deleted_at__isnull=True)
+        defect.deleted_at = timezone.now()
+        defect.save()
+
+        return Response(status=status.HTTP_204_NO_CONTENT)
